xinetd: removed ExecStop=
systemd will send SIGTERM to the PID stored in /var/run/xinetd.pid on stop. xinetd will then remove the pidfile and terminate.

There is therefore no need to delete the pidfile manually. Furthermore, if we did want to delete it, it should be done in ExecStopPost, as the way it is now systemd will not know what the main process is, so will not know where to send SIGTERM.